Refactor DagDashboard and DagProvider to use DagContext
- Move fetching logic, layout logic, and state management out of `DagDashboard` and into `DagProvider` inside `DagContext.tsx`.
- Refactor `DagDashboard` to consume state (`nodes`, `edges`, `isLoading`) directly from `useDagContext()`.
- Update `DagContext.tsx` types and `builder.ts` to properly pass `depends_on`.
- Update corresponding unit tests.
- Resolve type errors related to `depends_on` array casting.
- Fix all unit test issues matching expected layout formats.
