+++ src/utils/dag/builder.ts.orig
@@ -12,7 +12,6 @@ export interface GraphNode {
     status: string;
     owner_persona: string;
     rejection_count: number;
-    depends_on: string[];
   };
 }
 
@@ -54,7 +53,6 @@ export function buildDagGraph(parsedNodes: ParsedNode[]): DagGraph {
         status: node.data.status,
         owner_persona: node.data.owner_persona,
         rejection_count: node.data.rejection_count,
-        depends_on: node.data.depends_on,
       },
     });
   }
